requirements.txt
# Production dependencies — minimum strict
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
bcrypt==4.2.0
python-dotenv==1.0.1

# Production database (PostgreSQL) — optionnel, retirer pour SQLite seul
psycopg2-binary==2.9.9

# Serveur WSGI pour la production
gunicorn==23.0.0
